+++ plugins/apps/functions.go
@@ -154,40 +154,3 @@ func maybeCreateApp(appName string) error {
 		return createApp(appName)
 	})
 }
-
-func filterApps(apps []string) ([]string, error) {
-	if !common.PlugnTriggerExists("user-auth-app") {
-		return apps, nil
-	}
-
-	sshUser := os.Getenv("SSH_USER")
-	if sshUser == "" {
-		sshUser = os.Getenv("USER")
-	}
-
-	sshName := os.Getenv("SSH_NAME")
-	if sshName == "" {
-		sshName = "default"
-	}
-
-	args := append([]string{sshUser, sshName}, apps...)
-	b, _ := common.PlugnTriggerOutput("user-auth-app", args...)
-	filteredApps := strings.Split(strings.TrimSpace(string(b[:])), "\n")
-	filteredApps = removeEmptyEntries(filteredApps)
-
-	if len(filteredApps) == 0 {
-		return filteredApps, fmt.Errorf("You haven't deployed any applications yet")
-	}
-
-	return filteredApps, nil
-}
-
-func removeEmptyEntries(s []string) []string {
-	var r []string
-	for _, str := range s {
-		if str != "" {
-			r = append(r, str)
-		}
-	}
-	return r
-}
