+++ plugins/cron/subcommands.go
@@ -3,7 +3,6 @@ package cron
 import (
 	"fmt"
 
-	"github.com/dokku/dokku/plugins/apps"
 	"github.com/dokku/dokku/plugins/common"
 	"github.com/ryanuber/columnize"
 )
@@ -32,7 +31,7 @@ func CommandList(appName string) error {
 // CommandReport displays a cron report for one or more apps
 func CommandReport(appName string, format string, infoFlag string) error {
 	if len(appName) == 0 {
-		apps, err := apps.DokkuApps()
+		apps, err := common.DokkuApps()
 		if err != nil {
 			return err
 		}
