+++ plugins/network/subcommands.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"strings"
 
-	"github.com/dokku/dokku/plugins/apps"
 	"github.com/dokku/dokku/plugins/common"
 )
 
@@ -108,7 +107,7 @@ func CommandList() error {
 
 // CommandRebuildall rebuilds network settings for all apps
 func CommandRebuildall() error {
-	apps, err := apps.DokkuApps()
+	apps, err := common.DokkuApps()
 	if err != nil {
 		return err
 	}
@@ -126,7 +125,7 @@ func CommandRebuildall() error {
 // CommandReport displays a network report for one or more apps
 func CommandReport(appName string, format string, infoFlag string) error {
 	if len(appName) == 0 {
-		apps, err := apps.DokkuApps()
+		apps, err := common.DokkuApps()
 		if err != nil {
 			return err
 		}
