+++ plugins/proxy/subcommands.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"strings"
 
-	"github.com/dokku/dokku/plugins/apps"
 	"github.com/dokku/dokku/plugins/common"
 	"github.com/dokku/dokku/plugins/config"
 )
@@ -153,7 +152,7 @@ func CommandPortsSet(appName string, portMaps []string) error {
 // CommandReport displays a proxy report for one or more apps
 func CommandReport(appName string, format string, infoFlag string) error {
 	if len(appName) == 0 {
-		apps, err := apps.DokkuApps()
+		apps, err := common.DokkuApps()
 		if err != nil {
 			return err
 		}
