+++ plugins/resource/subcommands.go
@@ -1,7 +1,6 @@
 package resource
 
 import (
-	"github.com/dokku/dokku/plugins/apps"
 	"github.com/dokku/dokku/plugins/common"
 )
 
@@ -27,7 +26,7 @@ func CommandLimitClear(appName string, processType string) error {
 // CommandReport displays a resource report for one or more apps
 func CommandReport(appName string, format string, infoFlag string) error {
 	if len(appName) == 0 {
-		apps, err := apps.DokkuApps()
+		apps, err := common.DokkuApps()
 		if err != nil {
 			return err
 		}
