+++ plugins/scheduler/triggers.go
@@ -3,7 +3,6 @@ package scheduler
 import (
 	"fmt"
 
-	"github.com/dokku/dokku/plugins/apps"
 	"github.com/dokku/dokku/plugins/common"
 	"github.com/dokku/dokku/plugins/config"
 )
@@ -32,7 +31,7 @@ func TriggerInstall() error {
 		return fmt.Errorf("Unable to install the scheduler plugin: %s", err.Error())
 	}
 
-	apps, err := apps.DokkuApps()
+	apps, err := common.DokkuApps()
 	if err != nil {
 		return nil
 	}
